refactor(interview): #537 — extract scripts/interview_prep.py to findajob.interview package
Final M3 extraction — completes the 4/4 sequence (#539 triage, #541
notify, #542 prep, this).

Layout (per #538 — package shape with three small modules):
- findajob/interview/orchestrator.py (288 LOC) — main(), _generate(),
  _latest, _read_or_empty, notify; first orchestrator under the 300
  LOC soft cap because interview_prep was the smallest source (401)
- findajob/interview/role_runner.py (67 LOC) — run_role() (byte-
  equivalent dupe of findajob.prep.role_runner.run_role; cleanup PR
  consolidates)
- findajob/interview/sentinel.py (50 LOC) — SENTINEL_NAME,
  SENTINEL_STALE_AFTER_SECONDS, _sentinel_blocks_run

Module-load `load_env()` moved into main(), consistent with the prior
three extractions.

scripts/interview_prep.py shrinks 401 → 32 LOC.

Tests:
- New tests/test_interview_import_safety.py (5 tests) — load_env spy
  on orchestrator import; per-module smoke imports; SENTINEL_NAME +
  SENTINEL_STALE_AFTER_SECONDS constants preserved; and a bytecode-
  equivalence guard between findajob.interview.role_runner.run_role
  and findajob.prep.role_runner.run_role. The bytecode comparison is
  via __code__.co_code (ignores docstrings + comments + whitespace),
  paired with inspect.signature equality (catches param/type-hint
  changes). Fails if the two copies drift before the cleanup PR.
- tests/test_interview_prep_sentinel.py — drops the
  sys.path.insert(SCRIPTS) hack; imports findajob.interview.sentinel
  directly.
- tests/test_interview_prep_cost_logging.py — import path moved to
  findajob.interview.role_runner; patch target moved to
  findajob.interview.role_runner.complete.

CI green locally: 1865 passed, 1 skipped; ruff check + ruff format
--check both clean repo-wide; mypy clean on the new package.

After this PR lands, three duplications remain across the four
extracted packages — all preserved deliberately per the import-only
discipline:
- run_role() in findajob.{prep,interview}.role_runner
- notify() in findajob.{triage,prep,interview}.orchestrator (3 copies)
- abbrev_title() in findajob.prep.orchestrator + scripts/rename_folders.py

The cleanup PR (operator decides scope: M3's 6th child or M3+) folds
run_role + notify duplications into shared modules.
