+++ src/findajob/interview/__init__.py
@@ -1,0 +1,7 @@
+"""Interview-prep artifact generator — runs after the Interviewing transition.
+
+Entry point: scripts/interview_prep.py (a thin shim that imports from this package).
+Reads the existing prep folder (briefing, tailored resume, cover letter, optional
+recruiter critique) and EXPANDS the briefing's interview-questions and stories
+sections into a structured prep document.
+"""
